Add mouse cursor + Fix destroy building

+++ GamaUE5Interaction/Source/GamaUE5Interaction/People.cpp
@@ -78,7 +78,7 @@ int32 APeople::GetY()
 	return position.Y;
 }
 
-void APeople::SetPosition(int32 x, int32 y)
+void APeople::SetPosition(float x, float y)
 {
 	position.X = x;
 	position.Y = y;
